devcontainer: Add with adoc serve
Add the .devcontainer specification file.
This markup file is used to bring up GitHub Codespaces with the
necessary doc requirements installed and with the live editing daemon
running (adoc serve from adi_doctools).
Codespaces usage is billed to the user.
This is primarily intended for individuals who are unfamiliar with
Linux, seeking a plug-and-play solution.
The devcontainer specification is MIT licensed.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>